+++ .gitignore
@@ -7,4 +7,30 @@ Pascal/maths.o
 Pascal/maths.ppu
 Pascal/lang_en.ppu
 Pascal/lang_vi.ppu
-./app1cli*.deb
+C#/main/bin/Debug/net6.0/app1.exe
+C#/main/bin/Debug/net6.0/app1.deps.json
+C#/main/bin/Debug/net6.0/app1.runtimeconfig.json
+C#/main/bin/Debug/net6.0/app1.dll
+C#/main/bin/Debug/net6.0/ref/app1.dll
+C#/main/bin/Debug/net6.0/app1.pdb
+C#/main/obj/Debug/net6.0/app1.csproj.AssemblyReference.cache  
+C#/main/obj/Debug/net6.0/app1.GeneratedMSBuildEditorConfig.editorconfig
+C#/main/obj/Debug/net6.0/app1.AssemblyInfoInputs.cache
+C#/main/obj/Debug/net6.0/app1.AssemblyInfo.cs
+C#/main/obj/Debug/net6.0/app1.csproj.CoreCompileInputs.cache    
+C#/main/obj/Debug/net6.0/app1.dll
+C#/main/obj/Debug/net6.0/ref/app1.dll
+C#/main/obj/Debug/net6.0/app1.pdb
+C#/main/obj/Debug/net6.0/app1.genruntimeconfig.cache
+C#/utilsLib/bin/Debug/net6.0/utilsLib.deps.json
+C#/utilsLib/bin/Debug/net6.0/utilsLib.dll
+C#/utilsLib/bin/Debug/net6.0/ref/utilsLib.dll
+C#/utilsLib/bin/Debug/net6.0/utilsLib.pdb
+C#/utilsLib/obj/Debug/net6.0/utilsLib.csproj.AssemblyReference.cache
+C#/utilsLib/obj/Debug/net6.0/utilsLib.GeneratedMSBuildEditorConfig.editorconfig
+C#/utilsLib/obj/Debug/net6.0/utilsLib.AssemblyInfoInputs.cache  
+C#/utilsLib/obj/Debug/net6.0/utilsLib.AssemblyInfo.cs
+C#/utilsLib/obj/Debug/net6.0/utilsLib.csproj.CoreCompileInputs.cache
+C#/utilsLib/obj/Debug/net6.0/utilsLib.dll
+C#/utilsLib/obj/Debug/net6.0/ref/utilsLib.dll
+C#/utilsLib/obj/Debug/net6.0/utilsLib.pdb
