+++ C#/app1.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "app1", "main\app1.csproj", "{F013EEC2-58B0-4F04-8B18-E96AAF9A2A6B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "utilsLib", "utilsLib\utilsLib.csproj", "{095AEF17-F72C-4C21-BC28-255C86A18F8A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F013EEC2-58B0-4F04-8B18-E96AAF9A2A6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F013EEC2-58B0-4F04-8B18-E96AAF9A2A6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F013EEC2-58B0-4F04-8B18-E96AAF9A2A6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F013EEC2-58B0-4F04-8B18-E96AAF9A2A6B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{095AEF17-F72C-4C21-BC28-255C86A18F8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{095AEF17-F72C-4C21-BC28-255C86A18F8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{095AEF17-F72C-4C21-BC28-255C86A18F8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{095AEF17-F72C-4C21-BC28-255C86A18F8A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
