--- C#/bin/Debug/net5.0/app1.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v5.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v5.0": {
-      "app1/1.0.0": {
-        "runtime": {
-          "app1.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "app1/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
