--- C#/bin/Debug/net5.0/app1.runtimeconfig.dev.json
@@ -1,8 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\Le Bao Nguyen\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\Le Bao Nguyen\\.nuget\\packages"
-    ]
-  }
-}
