--- C#/bin/Debug/net5.0/app1.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "net5.0",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "5.0.0"
-    }
-  }
-}
