+++ debian/changelog
@@ -1,10 +1,10 @@
 app1cli (1.0.28-dev) unstable; ugency=medium
 
 * A Developer release of app1cli
-Make changes after a long time for C# edition
-Enable the "parse arguments" feature
-Add new math: x^2 + a^2
-Apply app1_utils library (my hand-made Pascal unit) for the program
+* Make changes after a long time for C# edition
+* Enable the "parse arguments" feature
+* Add new math: x^2 + a^2
+* Apply app1_utils library (my hand-made Pascal unit) for the program
 
 -- Le Bao Nguyen <bao12345yocoo@gmail.com> Fri, 21 Jan 2022 20:31 +0700
